+++ tests/ApiTests.php
@@ -11685,5 +11685,6 @@ require __DIR__ . '/Support/TransferLoadPlannerTests.php';
 require __DIR__ . '/Support/StorageBudgetTests.php';
 require __DIR__ . '/Support/GerminationDepotTests.php';
 require __DIR__ . '/Support/SectorStorageHttpTests.php';
+require __DIR__ . '/Support/OthersRepairTests.php';
 removeDirectory($tmp);
 exit($test->finish());